requirements.txt
streamlit
pandas
google-generativeai
pymysql
python-dateutil
requests